requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

#azure-functions
azure-functions

# Async Cosmos DB clie
azure-cosmos

# Async HTTP client
httpx

# RSS Feed Parser
feedparser
